# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "pos-tech",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost"
    },
    {
      "url": "localhost"
    }
  ],
  "paths": {
    "/produtos": {
      "post": {
        "tags": [
          "produtos"
        ],
        "summary": "cria produto",
        "description": "cria produto",
        "operationId": "cadastraProduto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoriaId": {
                    "type": "number",
                    "example": 3
                  },
                  "descricao": {
                    "type": "string",
                    "example": "descricao"
                  },
                  "imagem": {
                    "type": "string",
                    "example": "http://localhost:8080/img.jpeg"
                  },
                  "nome": {
                    "type": "string",
                    "example": "Lanchão"
                  },
                  "valor": {
                    "type": "string",
                    "example": "1000"
                  }
                }
              },
              "examples": {
                "cria produto": {
                  "value": {
                    "categoriaId": 3,
                    "descricao": "descricao",
                    "imagem": "http://localhost:8080/img.jpeg",
                    "nome": "Lanchão",
                    "valor": "1000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/clientes": {
      "post": {
        "tags": [
          "clientes"
        ],
        "summary": "cadastra cliente",
        "description": "cadastra cliente",
        "operationId": "cadastraCliente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "CPF": {
                    "type": "string",
                    "example": "44444444444"
                  },
                  "email": {
                    "type": "string",
                    "example": "teste2@gmail.com"
                  },
                  "nome": {
                    "type": "string",
                    "example": "Teste 2"
                  }
                }
              },
              "examples": {
                "cadastra cliente": {
                  "value": {
                    "CPF": "44444444444",
                    "email": "teste2@gmail.com",
                    "nome": "Teste 2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}